requirement.txt
transformers
datasets
huggingface_hub
safetensors
numpy
pandas 
scipy 
scikit-learn 
matplotlib
loguru
tqdm
h5py
json-repair
vllm
timm==1.0.20
peft
wandb
hydra-core
torchmetrics
openpyxl
deepspeed==0.17.1
pycatch22